linux: change the port selection algorithm in ftp alg
In the case of MAP feature, the ports are limited to number with same PSID,
the port should be in this format |offset|psid|pad|, the offset usually is
6 bits, the psid varies in 0~8 and pad length of the rest bits after offset
and psid.

the new port use same first ten bits of port after SNAT and the last 6 bits
of original port and varies the last 6 bits only.

this makes sure that new port and the port after snat share the same PSID.

Change-Id: I5e5a3c20fa0ec4e31273c5e23c6e6aa86ebb97f6
Signed-off-by: Ken Zhu <guigenz@codeaurora.org>