net sch: Add Qdisc flag for NSS
There is no check in nss qdisc clients to ensure if qdisc is allocated
by nss or not. It simply type cast the private structure of qdisc to nss
qdisc. If the private area of a non-nss qdisc is accessed and there is a
size mismatch, then it can cause out of boundary memory access. This is
what happens at boot time since by default qdisc is configured as
“noqueue” whose private size is 0. So there is a need to add flag in
qdisc structure to check the type of qdisc without accessing it's private
area.

Change-Id: I421b78b2868dd58689f5ea4c6523528ac2335763
Signed-off-by: Suruchi Agarwal <suruchia@codeaurora.org>
Signed-off-by: Tushar Ganatra <quic_tganatra@quicinc.com>