drivers:iio:stm:imu:st_lsm6dsox:Fix delay in read raw
When the raw data is read, the sensor is turned on, a delay equal to at
least one ODR is waited, and then the data is read from the register.
In this phase, when the ODR is set to 12.5 Hz the delay value is wrong
because uodr is expressed in us. The uodr value can be ignored since the
calculated delay is already higher than the configured ODR.

Signed-off-by: mario tesi <mario.tesi@st.com>
Change-Id: If21280bf46c943dd1e5883b5bfa7e5cc8a7fe263