drivers:iio:stm:imu:st_asm330lhhx: Add driver info management
Since st_asm330lhhx driver supports more devices, the recognition
of the device name has been included in the probe phase so that
the sensor name is referred to the device name. Discrimination
of the device has the purpose of verifying the presence or absence
of the mlc and the sensor hub features.

Signed-off-by: Mario Tesi <mario.tesi@st.com>
Change-Id: I0239d1b7d27be360f8918ec0db4b765fb543fb93
Reviewed-on: https://gerrit.st.com/c/linuxandroidopen/stm-ldd-iio/+/260651
Tested-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Denis CIOCCA <denis.ciocca@st.com>