drivers:iio:stm:imu:st_lsm6dsrx: manage multiple device id
Preparing the st_lsm6dsrx driver to extend support to devices of the
same family group.
The device ID managed through the compatible string has been introduced
in order to define which device to support through the device tree.
At the moment only lsm6dsrx which supports MLC is available.

Signed-off-by: Mario Tesi <mario.tesi@st.com>
Change-Id: Ib290fc4ffafab0a89fc171b26c6bddb5cab8ce95
Reviewed-on: https://gerrit.st.com/c/linuxandroidopen/stm-ldd-iio/+/264561
Tested-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Denis CIOCCA <denis.ciocca@st.com>