drivers:iio:stm:imu:st_lsm6dsrx: add tap embedded low-power features
LSM6DSRx family support tap and double tap detection embedded low-power
features. This patch configures the tap and double tap engine with a
basic configuration and enables the generation of interrupts in case of
acknowledgment of a tap or double tap event.
Each IIO event detection for tap and double tap can be individually enabled
as for the other basic embedded events.

Signed-off-by: Mario Tesi <mario.tesi@st.com>
Change-Id: Id0b2a6c149e18b266e90b1cc37cfced01187ba99
Reviewed-on: https://gerrit.st.com/c/linuxandroidopen/stm-ldd-iio/+/265161
Tested-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Denis CIOCCA <denis.ciocca@st.com>