drivers:iio:stm:imu:st_lsm6dsrx: add page lock during mlc load/flush
During the loading of the MLC / FSM fw the register page can be
changed randomly and repeatedly by the firmware itself, for this
reason it is necessary to protect the access to allow the MLC
event handler to read the status register page and output register
value in an exclusive mode.
Same possible issue for mlc flush operation which in turn invoke
the main sensor enable function that is not protected by any register
page switch.

Signed-off-by: Mario Tesi <mario.tesi@st.com>
Change-Id: I1146fbc3a58dfec1e241fd379e2d82d72f0d3861
Reviewed-on: https://gerrit.st.com/c/linuxandroidopen/stm-ldd-iio/+/265496
Tested-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Denis CIOCCA <denis.ciocca@st.com>