QCLINUX: dt-bindings: firmware: Add documentation for qcom,platform-parts-info
Add documentation to describe device tree bindings for QCOM's
platform-parts-info node. Firmware populates these nodes to pass the
information to kernel regarding the subset of hardware blocks
and features like Camera, Modem, Display present in a product.

This is to support that the same software image runs seamlessly on
different platforms where one or more HW blocks are not supported or
if some sub parts for a particular block are not supported.

Purpose of these definitions is to allow clients to know about this,
and thus, handle these cases gracefully.
For eg: Camera drivers and user space daemons can use this to know
if camera is not supported on the device, or some particular HW blocks
inside a camera are not supported.

Firmware also populates board and product related information like
memory card's serial number and SKU(Stock Keeping Unit) details like
product code and feature code for clients to consume.

Change-Id: I26765ff97e1cffc84f6f0158d4980cf27bb10132
Signed-off-by: Naman Jain <quic_namajain@quicinc.com>