drivers:iio:stm:imu:st_lsm6dso16is: add lsm6dso16is device to kernel build
The st_lsm6dso16is driver was not included in the kernel build
because it was not added in the top level Kconfig/Makefile,
furthermore the default config contained a typo.

In lsm6dso16is_defconfig there is a wrong device name,
this fix enables lsm6dso16is module compilation.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: I92df0ca4c3e29af6bcaef5ce80a78d494672ef00
Reviewed-on: https://gerrit.st.com/c/linuxandroidopen/stm-ldd-iio/+/325964
Reviewed-by: Matteo DAMENO <matteo.dameno@st.com>
Tested-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>