QCLINUX: soc: qcom: Add support for watchdog framework
Add support for watchdog framework. Provides interface to register
a watchdog device and implements a wdog that checks for cpu health
with ipi calls to cores before resetting the watchdog, resets system
in case of a core stuck, registers panic notifier to call bite on
panic and implements other wdog crash debug features.

Change-Id: Ic5ef5a7f9c5499991fcc19ebdddc25a4701dd687
Signed-off-by: Shivendra Pratap <quic_spratap@quicinc.com>