drivers:iio:stm:accel:st_lis2dw12: Add multiple interrupt management
The sensor family whose founder is lis2dw12 includes the possibility
to internally detect some gestures such as TAP, double TAP, wake up,
etc.

These events are notified to the system processor via a dedicated
interrupt pin which in the previous version of the driver could not
be dissociated from the accelerometer sensor interrupt pin.

With this new version of the driver it is now possible to define a
dedicated interrupt for the management of embedded events distinct
from the accelerometer sensor interrupt.

Signed-off-by: Mario Tesi <mario.tesi@st.com>
Change-Id: Iac61300266582d01c0505605189c874abd1f6eb0
Reviewed-on: https://gerrit.st.com/c/linuxandroidopen/stm-ldd-iio/+/339390
Reviewed-by: Matteo DAMENO <matteo.dameno@st.com>
Tested-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>