PENDING: phy: qcom-qmp-pcie: Add support for keeping refclk always on
In PCIe low power states like L1.1 or L1.2 the phy will stop
supplying refclk to endpoint. If endpoint asserts clkreq to bring
back link L0, then RC needs to provide refclk to endpoint.

Some platforms with pcie switch fail to drive the clkreq signal to
the host from the endpoints because of the switch board design.
Due to that refclk needs to supplied to the endpoint always.

Add a flag to keep refclk always supplied to endpoint.

Change-Id: I7aaaa25589980ab38eb0fcc7a617c644ca8c656b
Signed-off-by: Krishna chaitanya chundru <quic_krichai@quicinc.com>