drivers:iio:stm:imu:st_asm330lhhx: disable irq line during suspend
The IRQ enable/disable status is independent of the IRQ enable
status so an IRQ can be disabled and still wake up the system
as long as the IRQ has wake feature active.
For this reason, while system is suspending, it is more correct
to disable the interrupts during the suspend phase and enable
them again at the end of the resume phase.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: Ib90cb491f93490b35c7416898eeac9903472c97d