sched/walt: Add AMU sysreg read checks gated by CONFIG_ARM64_AMU_EXT
When reading sysreg on AMU, we should not only check the enablement of
the configuration but also validate the AMU implementation on the target
CPU. Implement this in such a way that the additional check is performed
only when CONFIG_ARM64_AMU_EXT is enabled. This caters to targets that
enable the config (due to SPF mechanism) but still do not support the
counters.

Note : This implementation only checks one core and does not support
heterogeneous architectures where cores differ in ISA levels (e.g., some
cores are ARMv8.2 and others are ARMv9). The current design does not
aggregate feature capabilities across mixed-architecture systems.

Change-Id: Ifa00211c5aa673e83a300cdbbb7c346ed667de7d
Signed-off-by: Song Xue <quic_songxue@quicinc.com>
Signed-off-by: Ankit Sharma <ankit.sharma@oss.qualcomm.com>
(cherry picked from commit 0bd484923f72189800ef967284a2c1351650a1cb)