drivers: soc: qcom: add snapshot of system sleep activities driver
When all the CPUs in the Apps processor subsystem
are powered off, the system may also reduce its
bandwidth and shared resource requirements.

The resource requests are sent through the SMD
from APSS subsystem. The sleep requests also need
to be written to the rpm to save power. This driver
handles the last few steps when entering system sleep.

When notified, the next wakeup timer value is written
to the MPM to be setup in the always-on domain timer.

System sleep activities driver snapshot from msm-5.4 commit a36917be8131
("msm: kgsl: Do not free sharedmem if it cannot be unmapped").

Change-Id: I1c26b5190a933b3b34e2e2d41622e0e6bbfb1cea
Signed-off-by: Sudharshan Reddy Yettapu <quic_syettapu@quicinc.com>