USB: dwc3: Adding GPIO-based regulator functionality for VBUS
The vbus_dwc3-supply property in the device tree manages
the 5V for USB host mode.

It is designed for platforms with hardware VBUS detection
based on a GPIO regulator, allowing the system to control
the USB port power supply, enabling or disabling
it in host start and stop cases.

Change-Id: I38c05670aaf0460d5781320641f388a6e261d479
Signed-off-by: Rajashekar kuruva <quic_kuruva@quicinc.com>