BACKPORT: wifi: cfg80211/mac80211: move puncturing into chandef
Aloka originally suggested that puncturing should be part of
the chandef, so that it's treated correctly. At the time, I
disagreed and it ended up not part of the chandef, but I've
now realized that this was wrong. Even for clients, the RX,
and perhaps more importantly, CCA configuration needs to take
puncturing into account.

Move puncturing into the chandef, and adjust all the code
accordingly. Also add a few tests for puncturing in chandef
compatibility checking.

Git-Commit: b82730bf57b54803ab94abbfd8c4422a7081886d
Git-Repo:
https://git.kernel.org/pub/scm/linux/kernel/git/wireless/wireless-next.git
Suggested-by: Aloka Dixit <quic_alokad@quicinc.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
CRs-Fixed: 4219094
Change-Id: I50f65f3e0cfdc4277131f4182c7a363f09ee7c5d
[mpaluri: Fix minor conflicts in chan.c, mlme.c, util.c, cfg.c]
Signed-off-by: Manaswini Paluri <quic_mpaluri@quicinc.com>
Signed-off-by: Huang Chenming <chenhuan@qti.qualcomm.com>