FROMLIST: PCI: qcom: reshuffle reset logic in 2_7_0 .init
At least on SC8280XP, if the PCIe reset is asserted, the corresponding
AUX_CLK will be stuck at 'off'. This has not been an issue so far,
since the reset is both left de-asserted by the previous boot stages
and the driver only toggles it briefly in .init.

As part of the upcoming suspend prodecure however, the reset will be
held asserted.

Assert the reset (which may end up being a NOP in some cases) and
de-assert it back *before* turning on the clocks in preparation for
introducing RC powerdown and reinitialization.

Link: https://lore.kernel.org/linux-pci/20240210-topic-8280_pcie-v2-1-1cef4b606883@linaro.org/
Change-Id: If7d3a34ad9307fb9459abf83bf5d6a196a4f6c9e
Patch-mainline: linux-pci@vger.kernel.org @ 10/02/2024, 17:10
Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>
Signed-off-by: Krishna chaitanya chundru <quic_krichai@quicinc.com>