FROMLIST: PCI: qcom: properly implement RC shutdown/power up
Currently, we've only been minimizing the power draw while keeping the
RC up at all times. This is suboptimal, as it draws a whole lot of power
and prevents the SoC from power collapsing.

Implement full shutdown and re-initialization to allow for powering off
the controller.

This is mainly intended for SC8280XP with a broken power rail setup,
which requires a full RC shutdown/reinit in order to reach SoC-wide
power collapse, but sleeping is generally better than not sleeping and
less destructive suspend can be implemented later for platforms that
support it.

Link: https://lore.kernel.org/linux-pci/20240210-topic-8280_pcie-v2-3-1cef4b606883@linaro.org/
Change-Id: Id0d2c842ff378ef84f7d7248f39714e0c240a61d
Co-developed-by: Bjorn Andersson <quic_bjorande@quicinc.com>
Patch-mainline: linux-pci@vger.kernel.org @ 10/02/2024, 17:10
Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>
Signed-off-by: Krishna chaitanya chundru <quic_krichai@quicinc.com>