drivers:iio:stm:accel:st_lis2du12: Rework event handler for fixing reset of latched interrupt
When the interrupt pin are shared between multiple sources it may
happen that reading the ALL_INT_SRC register causes the reset of
the interrupt status bits. To avoid this you must try to read
ALL_INT_SRC only when one of the IA bits is set in the status
registers (TAP_SRC, WAKE_UP_SRC and SIXD_SRC).

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: I21c855ae8929163bd550c269165ae83a87718864