FROMLIST: dt-bindings: Add binding for gunyah hypervisor
The Gunyah Resource Manager applies a devicetree overlay describing the
virtual platform configuration of the guest VM, such as the message
queue capability IDs for communicating with the Resource Manager. This
information is not otherwise discoverable by a VM: the Gunyah hypervisor
core does not provide a direct interface to discover capability IDs nor
a way to communicate with RM without having already known the
corresponding message queue capability ID. Add the DT bindings that
Gunyah adheres for the hypervisor node and message queues.

Change-Id: I6a6b7cd3e669cbae622954dcb965fc13ea8ceda0
Reviewed-by: Rob Herring <robh@kernel.org>
Signed-off-by: Elliot Berman <quic_eberman@quicinc.com>
Link: https://lore.kernel.org/all/20240222-gunyah-v17-2-1e9da6763d38@quicinc.com/
Signed-off-by: Sayan Dey <quic_sayand@quicinc.com>