PENDING: soc: qcom: f_qdss: Add support for QDSS function driver
QDSS (Qualcomm Debug Sub-System) provides faster debug functions,
merged interleaved trace, and application SW trace messaging.
Various ARM CoreSight IP blocks are utilized in QDSS because of
their proven multi-core system debug and trace capabilities.

From the kernel point of view, this is a composite driver with only
one in endpoint which is used to send data out of the device.

Following is the functionality with USB:

* Trace data will be collected from multiple trace sources
on different processors and sent to the host over various
trace links.
* With QDSS, there is no need of USB RX Functionality
(i.e. No USB-OUT data transfer ), and only USB-IN data
transfer is required to send received qdss trace information.

The driver coresight is the one who collects all the data from
the various sources and chooses which channel to use for sending
this data. When the coresight chooses usb as the medium to transfer
then the usb software path will be established with one BULK IN
endpoint.

Coresight driver opens/closes the channel for usb via following API:
* usb_qdss_open: Accepts channel name, driver_data, and a notifier.
* usb_qdss_close: Accepts the existing usb channel.

With the passed notifier to the usb_qdss_open function, f_qdss &
coresight further communicates.

Further, coresight driver uses write API to transfer data:
* usb_qdss_write: Accepts usb channel, usb_request.

As per the gadget frameworks the usb requests when completes
generates a complete callback which is registered back in usb_qdss_open
* usb_qdss_write_complete: Further notifies the coresight of completion
			   of a particular usb request.
* usb_qdss_free_req: Accepts usb channel and frees all the requests.

Change-Id: I02165cc95f6daa8b509f70ac35a0f43f17b0d7fb
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>