QCLINUX: drivers: soc: qcom: Add support of KVM enablement
Ideally there should be no need to distinguish between the Gunyah and
KVM hypervisors in drivers/dT during the build process. However, there
are exceptions, such as watchdog and PIL loading, which require device
tree-level modifications based on the hypervisor. Since the device tree
cannot detect Kbuild configs, add QCOM_KVM_ENABLED config which is used
to pass a compile time option to dT compiler.
The corresponding device tree nodes will utilize this compile-time flag
to include conditional code (#ifndef) specific to the hypervisor.
This compile-time framework may be revised in the future if a unified
solution for watchdog and PIL loading is identified.

Change-Id: I6e2f9dec63c4546c2a74075b97bf484ca508547c
Signed-off-by: Mukesh Pilaniya <quic_mpilaniy@quicinc.com>