PENDING: smem: Register system wake-up interrupt line
Glink interrupt registered with IRQF_NO_SUSPEND flag does
not guarantee that this interrupt will wake the system from
a suspended state. For USB insertion in suspended system
interrupts ADSP and ADSP in turn sends glink interrupt to
APSS to wake up the system. But glink interrupt is not able
to wake up the APSS in this scenario.

Removing IRQF_NO_SUSPEND flag and marking glink interrupt with
enable_irq_wake to make it wakeup capable.

Change-Id: Idad079430e7d0661367df54d4a33a9820f3f8526
Signed-off-by: Kaushal Hooda <quic_khooda@quicinc.com>