power: supply: qti_battery_charger: Detect second USB dynamically
In the current design, second USB port support is enabled based on a
DT property alone. On the other hand, the charger firmware that runs on
the remote processor detects the second port presence dynamically
reading a HW fuse.

To make the design seamless, charger firmware has added a new
property under first USB port, to provide the number of USB ports
supported.

Add logic to read this new property, in addition to the existing DT
property based detection, to enable support for the second port.

Change-Id: Ie70294cb314c6b53e60c36398665ae4f218ea662
Signed-off-by: Jishnu Prakash <quic_jprakash@quicinc.com>