drivers:iio:stm:imu:st_asm330lhhx: rework timesync to guarantee a minimum number of syncs for all sensors
With this patch we guarantee at least 10 fast sync samples at
each timesync instance to have a good estimate of the sensor
timestamp.
In the past the number of fast samples was not guaranteed to
each sensor so it could happen that the fast sync samples ran
out too soon before the sensor even started sending data.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: Icada143e15cc92129f6e0867d25d1458aad8f3dd