drivers:iio:stm:imu: add the possibility to exclude custom event management from build
The introduction of IIO custom event detection like free fall,
wake-up, step counter etc. leads to the registration of IIO
types on the iio core in the kernels that do not have a valid
index in the iio_chan_type_name_spec array which could lead to
a kernel index out of range event, for this reason the possibility
of excluding support for custom events was introduced in Kconfig
of IMU STMEMS device drivers.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: I3e7d112516d3697fba0e3ab412b54b8816bd503e