drivers:iio:stm:accel:st_lis2duxs12: fix power-up sequence on SPI bus
If the SPI interface is used, the LIS2DUXS12 can move from
deep power-down to soft power-down by writing the bit SOFT_PD
of EN_DEVICE_CONFIG (3Eh) to 1. In order to verify that device
has correctly completed the transition to soft power-down, the
who_am_I can be checked. This procedure is not applicable when
I2C bus interface is used.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: Ib60cf723e1af7813aeac9f1fad3536b061737364