PENDING: crypto: qce - Bam crash during suspend-shutdown scenario
In the current implementation of qce crypto driver, during
driver_suspend the clocks/icc are disabled and in a scenario
where driver_suspend call is followed by driver_shutdown the
bam registers are accessed. Since the clocks/icc is in reset
state due to driver_suspend call the access violation is observed
resulting in system crash.

To fix this set/reset the interconnect during shutdown sequence.

Test: crypto usecases and suspend/shutdown scenario validation.

Change-Id: I37ee301d062b5f668afeac26af4e801e49ad11f3
Signed-off-by: Udit Tiwari <quic_utiwari@quicinc.com>