drivers:iio:stm:imu:st_lsm6dsvx: optimize wait time in I2C master
To perform operations on the I2C master, it is necessary to allot
sufficient time to perform it, based on the fact that each operation
requires at least one period equal to the current ODR of the
primary sensor, which in this case is the accelerometer.
With this patch, we start allowing at least one ODR + 10% before
verifying if the operation has indeed been completed, to ensure
that the I2C master transaction has been truly finalized.
Indeed, since the I2C isn't really a fast bus, in addition to the
ODR that starts the transaction, enough time must elapse for the
operation to be fully concluded.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: If5f7020c4e5a13b7e24c18eee8ea3d72282b5252