drivers:iio:stm:imu:st_ism330dhcx: optimize wait time in I2C master
To perform operations on the I2C master, it is necessary to allot
sufficient time to perform it, based on the fact that each operation
requires at least one period equal to the current ODR of the
primary sensor, which in this case is the accelerometer.
With this patch, we start allowing at least one ODR + 20% before
verifying if the operation has indeed been completed, to ensure
that the I2C master transaction has been truly finalized.
Indeed, since I2C is not exactly a fast bus, especially if as in
this case the master I2C clock is at 100 kHz, in addition to the
ODR that starts the transaction, a sufficient amount of time must
pass for the operation to be completely concluded.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: Id31b1bee89881924b36533ab133ade7c585285ef