drivers:iio:stm:imu:st_lsm6dsvx: rework embedded function and events interrupt management
The embedded functions and event detection requires use
of hardware interrupts pin configured and available, so
at least one of the two interrupt pins of the sensor must
be configured.

Before this patch there was no check to see if irq pin was
actually configured in device tree, causing a malfunction
if irq line was not used, e.g. when SW triggers were
enabled.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: I942c63af94ca60b471c437e5cb0e5ce29eb0b7a3