PENDING: can: mcp251xfd: enable CAN transceiver using gpio_0
The GPIO_0 on mcp251xfd chip can be configured as an additional
interrupt for RX frames as well as a general-purpose Input/Output pin.
The GPIO_0 is connected to a CAN transceiver, which has 2 modes of
operation - normal mode and standby mode. The CAN frames are transmitted
when the transceiver is in normal mode. When gpio_0 is LOW the
transceiver is in normal mode and is in standby mode then gpio_0 is
high.

The mode of transceiver is set using the function -
mcp251xfd_transceiver_mode(), which sets the transceiver in normal mode
when mcp251xfd device is opened and sets the transceiver in standby mode
when the mcp251xfd device is closed.

Change-Id: I0dea62a0643c856050808286a7b972f9961bd1c3
Signed-off-by: Anup Kulkarni <quic_anupkulk@quicinc.com>