pci: msm: Use of_address_to_resource to parse msi_addr
Currently msi_addr is parsed from devicetree by obtaining the location
of the reg property and directly parsing the value of the first cell.
This may work when address-cells is fixed at 1, but will be parsed
incorrectly if the address-cells is 2 where the first cell will only be
the upper 32-bits of the address.

Additionally, the current parsing mechanism uses the direct value of the
cell without any translation. This could be an issue if parent bus do
not provide a direct mapping.

To address these issues, use of_address_to_resource() to parse the reg
property, which takes into account both address-cells and address
translations.

Change-Id: Ic071e1342cc822ae92b9d535c851779e585aa981
Signed-off-by: Matthew Leung <quic_mattleun@quicinc.com>
Signed-off-by: Sai Chaitanya Kaveti <quic_skaveti@quicinc.com>