drivers:iio:stm:accel:st_lis2dw12: fix wakeup management
The sensor driver can handle the wakeup source only if it
is the same interrupt pin as the data irq.

This patch allows to diversify the interrupt sources.
Considering that wakeup irq is available only on int pin 1
by hardware, the possible configurations will be:
- wake-up and data irq on int pin 1
- wake-up irq on int pin 1 and data irq on int pin 2

In the first case, only one irq configured in the device
tree will be enough while in the second case, it will be
necessary to describe two distinct vectors in the device
tree where the one with the lowest index will be related
to the data irq while the one with the highest index will
be that of the system wake-up

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: Id3eccf8498cf472cd0a5923e0c412d3583c55139