QCLINUX: soc: qcom: smem: Switch partitions to xarray
The partitions array is statically sized and uses the remote host ID as
its index. There are valid remote host IDs which greatly exceeds the
current defined size. Instead of increasing the size of the array and
wasting memory for large remote host IDs, switch over to using xarray.
This provides partitions xarray to grow dynamically and save memory.

Change-Id: Ia2c80bdb5bfd06ae7c91b3eebb27d21d27fd6762
Signed-off-by: Tony Truong <quic_truong@quicinc.com>