defconfig: sdxkova: Enable NLS configs needed for VFAT for sdxkova
Enable NLS configs needed for VFAT for sdxkova.
NLS stands for Native Language support. These configs are crucial
for enabling support for various character sets and encodings, which
is essential for filesystems that store filenames in different
languages. So this is needed for multilingual support and to use
different encoding schemes for filenames.

Change-Id: I07d3c67b2cef7e035f41be434133a98b4e86678c
Signed-off-by: Sarthak Garg <quic_sartgarg@quicinc.com>