net: qrtr: Add support for a mhi device transport
Add a transport that uses the mhi device APIs. The mhi device APIs
are used to transfer the data from a device to the host over MHI.
The modem will generally be configured as a device and this mhi
transport is intended to be used on the modem controller co-processor.

Add support for mhi device transport driver from msm-5.15.c1 branch
commit 737cf08fe27a ("Merge "Revert "msm: ipa: Add support for Vlan
Priority""").

Change-Id: Ie34f47482c9a45fb15b2c49014b8856212d1812f
Signed-off-by: Vishnu Santhosh <quic_vishsant@quicinc.com>