net: Add extra priv_flag support in netdevice
This patch adds a flag in struct net_device which
is an extension to priv_flags and is used to
check the type of device.

Add GRE std support in device tree and priv_flags.

Added dts support and priv flags support for GRE and
to check the type of device.

Change-Id: Ibb5a3f0c706e1b169b848aed7004aa4ea6d57069
Signed-off-by: SHASHANKA SHEKHER NAYAK <quic_shasnaya@quicinc.com>