ppp: PPPoE acceleration support
pppoe: Add return value to pppoe addressing get function

If the addressing doesn't have a netdevice, this should be
handled as a failure, so that the caller considers that
the get function failed.

Added some new APIs to the PPP/PPPoE kernel modules
for using from the hardware acceleration connection managers.

Change-Id: I2c16c6d6ccba8ffa14aec077c8dad1681535ae0b
Signed-off-by: SHASHANKA SHEKHER NAYAK <quic_shasnaya@quicinc.com>