serial: msm_geni_serial: Return error in probe upon memory allocation failure
When allocating memory for rx_fifo using devm_kzalloc, there’s a
possibility that the allocation might fail due to insufficient memory.
If this happens and the failure is not handled, it could lead to
dereferencing a NULL pointer.

By adding these checks, ensured that the code handles memory allocation
failures properly, preventing potential crashes or undefined behavior
in the kernel.

Change-Id: Ic1d784dca2ffeb3aeaa71879d5538294df9de9ca
Signed-off-by: Somesh Dey <quic_somedey@quicinc.com>