cfg80211: Add support for ML STA fallback association
In 6.6 kernel, when MLO STA downgrades itself to connect to legacy AP,
the connect response from driver does not contain any wdev->valid_links
information since the association is a legacy association.

To support this, add fallback_valid_links bitmap as part of connect
response as well as wdev to identify the link that has downgraded and
use the same for deletion of the link and other similar operations.

Change-Id: I04316e7861b75b173e5752afff7d928a56db9561
Signed-off-by: Abishek Ganapathy <quic_abishekg@quicinc.com>