PENDING: soc: qcom: qup_fw: QUP SE Firmware load from Linux
Provide an interface to load QUP SE firmware from the Linux.
Populate SE details and base address details in the probe
function of the protocol driver and pass to firmware load
routine. Load the firmware to QUP SE if the "load-firmware"
property is specified in dtsi. Skip the firmware loading
process if the firmware is already loaded in SE.

Change-Id: I1b460a5a1851ddafa8a373e639498815d20addfb
Signed-off-by: Mehul Raninga <quic_mraninga@quicinc.com>