BACKPORT: FROMLIST: dm-inlinecrypt: add target for inline block device encryption
Add a new device-mapper target "dm-inlinecrypt" that is similar to
dm-crypt but uses the blk-crypto API instead of the regular crypto API.
This allows it to take advantage of inline encryption hardware such as
that commonly built into UFS host controllers.

The table syntax matches dm-crypt's, but for now only a stripped-down
set of parameters is supported.  For example, for now AES-256-XTS is the
only supported cipher.

dm-inlinecrypt is based on Android's dm-default-key with the
controversial passthrough support removed.  Note that due to the removal
of passthrough support, use of dm-inlinecrypt in combination with
fscrypt causes double encryption of file contents (similar to dm-crypt +
fscrypt), with the fscrypt layer not being able to use the inline
encryption hardware.  This makes dm-inlinecrypt unusable on systems such
as Android that use fscrypt and where a more optimized approach is
needed.  It is however suitable as a replacement for dm-crypt.

Change-Id: I49261a1514d6a8431c0c65d03ead1cade06ec0db
Link: https://lore.kernel.org/r/20241016232748.134211-3-ebiggers@kernel.org/
Link: https://lore.kernel.org/all/20260304121729.1532469-1-linlin.zhang@oss.qualcomm.com/
Signed-off-by: Eric Biggers <ebiggers@google.com>
Patch-mainline: dm-devel @ 16/10/2024, 16:27
[linlin.zhang@oss.qualcomm.com: Abort map I/O when BIO
 hasn't any data and add hardware wrapped key support]
Signed-off-by: Linlin Zhang <linlin.zhang@oss.qualcomm.com>