PENDING: Bluetooth: hci_sync: Fix disconnect complete event timeout issue
Sometimes the remote device doesn't acknowledge the
LL_TERMINATE_IND in time, requiring the controller to wait
for the supervision timeout, which may exceed 2 seconds.
In the current implementation, the HCI_EV_DISCONN_COMPLETE
event is ignored if it arrives late, causing the mgmt to
get stuck. As a result, bluetoothd waits indefinitely for
the mgmt response to the disconnect.

bluetoothctl log like this:
[Designer Mouse]# disconnect D9:B5:6C:F2:51:91
Attempting to disconnect from D9:B5:6C:F2:51:91
[Designer Mouse]#
[Designer Mouse]# power off
[Designer Mouse]#
Failed to set power off: org.freedesktop.DBus.Error.NoReply.

Change-Id: Iee64f356b1dfa0c69dd2d269ce4592144181dd40
Signed-off-by: Cheng Jiang <quic_chejiang@quicinc.com>