USB: gadget: Add extra allocation size to use APF feature for IN eps
APF feature requires the software to allocate additional 256
bytes for TX/IN buffers. This is required because optimized
pre-fetching makes hardware to read outside the allocated buffer.

To protect out of boundary access, SW needs to allocate an
extra 128 bytes or more to make sure the neighbouring region
is not protected by TZ or MSA. Introduce an extra_alloc_buf
in usb gadget which can indicate this extra allocation. Enable
this only if APF is enabled.

Change-Id: I4632ce2b6a4ecd95bc0c9e787ca57de8a9088b57
Signed-off-by: Sandeep Maheswaram <quic_sandmahe@quicinc.com>