Revert "Shortcut-fe: using existing api to avoid gki issue"
This reverts commit 5cf1ca07969a7d917e5db5a4ade2337dec347571.

Reason for revert: The `br_port_dev_get` API was updated to use the
existing `br_fdb_find_port` API in the userspace SFE driver. This
change eliminated the need for additional file modifications to
maintain mainline compatibility. However, it introduced a regression
where WLAN peer teardown was not happening for connected clients
using the QCMAP CLI application.

To resolve the issue, the changes were reverted to restore functionality.

Change-Id: Ia05787bad5e3dfdfb6ed4055abc9583ea7f59971
Signed-off-by: Mandar Mahesh Kamble <quic_maheshka@quicinc.com>