PENDING: usb: dwc3: drd: expose role-switch control to userspace
For the Carplay usecase which is supported by the Apple Accessory
Application (AAA), when the Device Under Test (DUT) boots up,
it initially assumes the USB host role, meaning it can control
other devices connected to it. When an iPhone is connected to the
DUT, the iPhone is recognized and enumerated as a device.

If the user decides to enable CarPlay, the AAA sends specific
commands to the DUT to request a switch in USB roles. This switch
is crucial because, for CarPlay to function correctly, the iPhone
needs to act as the host, while the DUT must switch to the device
role. This role switch request is managed by the DUT, allowing the
user space entity (software running on the DUT) to change the USB
role from host to device.

During this transition, the iPhone will temporarily disconnect from
the USB bus, switch its role to host, and expect the DUT to switch
to the device role. After the role switch is completed, the iPhone
will act as the host, controlling the connection, while the DUT
will function as the device, being controlled by the iPhone.

The allow_userspace_control property enables manual switching of USB
roles from the user space via sysfs entry nodes. This manual control
can be utilized by the AAA to ensure the role switch occurs as needed
for CarPlay functionality.

Add the userspace control functionality as a non-DT property.

Change-Id: I1c531ed2bac8a98dbd50e98c409424bdc20e808c
Signed-off-by: Sriram Dash <quic_sriramd@quicinc.com>