usb: dwc3: qcom: Remove ACPI support from glue driver
Minimal ACPI support was added to the Qualcomm DWC3 glue driver in order to
enable USB on SDM850 and SC8180X compute platforms. The support is still
functional, but unnoticed regressions in other drivers indicates that no
one actually booting any of platforms dependent on this implementation.

The functionality provides is the bare minimum and is not expected to aid
in the effort of bringing full ACPI support to the driver in the future.

Remove the ACPI code from the Qualcomm DWC3 glue driver to aid in the
implementation of improvements that are actually used like multiport and
flattening device tree.

Change-Id: I5dbce163fa910a70debd1c52335f71a0a9fc2cd8
Signed-off-by: Rajashekar kuruva <quic_kuruva@quicinc.com>
(cherry picked from commit 4448b51b6afcf99bc12b4acfc614420672a230c3)
Signed-off-by: Jayasri Bhattacharyya <quic_jbhattac@quicinc.com>