usb: dwc3: msm: Allow prepare/complete callbacks for only msm-dwc3 child
In the dwc3-msm-core, core USB PM operations are
overridden with prepare and complete callbacks for
both host and peripheral modes.

To address an issue with the MP controller,
which uses the QCOM driver and lacks callback
implementations, a conditional check has been added
in the prepare and complete functions to avoid
callbacks for controller that don't use msm-driver.

Change-Id: Ibd03b78eef9c176dc0c8e82ef6723f767a5ac5b1
Signed-off-by: Rajashekar kuruva <quic_kuruva@quicinc.com>