usb: dwc3: msm: Prevent prepare/complete callbacks for non-MSM core drivers
In multicontroller scenarios for PM operations, the last controller
to update the operations receives callbacks for all controllers.

This issue is observed with the MP controller,
which uses the QCOM driver and lacks callback implementations.

To address this, add a conditional check in the prepare and
complete functions to avoid callbacks for the MP controller.

Change-Id: Ibd03b78eef9c176dc0c8e82ef6723f767a5ac5b1
Signed-off-by: Rajashekar kuruva <quic_kuruva@quicinc.com>
(cherry picked from commit 20e1b2aed3005751061f7427f0b235b838656444)
Signed-off-by: Jayasri Bhattacharyya <quic_jbhattac@quicinc.com>