FROMLIST: iommu/arm-smmu: introduction of ACTLR for custom prefetcher settings
Currently in Qualcomm  SoCs the default prefetch is set to 1 which allows
the TLB to fetch just the next page table. MMU-500 features ACTLR
register which is implementation defined and is used for Qualcomm SoCs
to have a custom prefetch setting enabling TLB to prefetch the next set
of page tables accordingly allowing for faster translations.

ACTLR value is unique for each SMR (Stream matching register) and stored
in a pre-populated table. This value is set to the register during
context bank initialisation.

Change-Id: Icb4ca116a39b1559bfe8968e14d0192a740b2e82
Link: https://lore.kernel.org/all/20241212151402.159102-5-quic_bibekkum@quicinc.com/
Patch-mainline: linux-arm-msm @ Thu, 12 Dec 2024 20:44:01 +0530
Reviewed-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
Signed-off-by: Bibek Kumar Patro <quic_bibekkum@quicinc.com>