qcom_scm: Allow reusing used SMC message buffer
The driver assumed the message buffer is unused and therefore the sequence
number in the mailboxes is zero, set the next expected response id to 1.

Read the both the request and response id from the buffer.
Check if all the previous requests are finished and throw an error if not.
Set the expected response to be one higher then the last request.

Change-Id: Ic2bb19496a8de548b1f946d3fe0048a2d584fa34
Signed-off-by: Robert Kuban <rkuban@qti.qualcomm.com>
Signed-off-by: Rafal Skierkowski <rskierko@qti.qualcomm.com>
(cherry picked from commit b048847ad109abadff438b3ca8bc43e3570c41aa)