usb: dwc3: msm: Add Multiport support for USB controller
Disable the pm overriding ops for child as child is used
simultaneously by 2 parent drivers. This way, avoid
layering violation from qcom to mdwc.

For device mode, if core complete is not called, then
the resume work is not queued, which doesnt switchs the
gdsc and clks ON.

This way, the controller gets soft reset fails.

Avoid by force resume for device mode to switch clocks
and gdsc ON.

Change-Id: Ic87230b6d20288813423cea6f4105646bea8ab2b
Signed-off-by: Sriram Dash <quic_sriramd@quicinc.com>