FROMLIST: dt-bindings: net: stmmac: Tune rx sampling occasion
Add documentation detailing the capability for tuning the RX sampling
occasion of the Qualcomm MAC. Introduce a new flag to indicate that a
board requires switching the timing of signal sampling, instead of relying
solely on the EMAC version in the driver code. For older DTS, the logic
for current boards is hard-coded in the driver code to ensure they
function as previously.

Link: https://lore.kernel.org/all/20241225-support_10m100m-v1-1-4b52ef48b488@quicinc.com/
Change-Id: I85e22ff77edbc4b2d403de3267c7bd4e454718cb
Signed-off-by: Yijie Yang <quic_yijiyang@quicinc.com>