BACKPORT: media: venus: Add support for static video encoder/decoder :declarations
Add resource structure data and probe() logic to support static
declarations of encoder and decoder.

Right now we rely on video encoder/decoder selection happening in the dtb
but, this goes against the remit of device tree which is supposed to
describe hardware, not select functional logic in Linux drivers.

Provide two strings in the venus resource structure enc_nodename and
dec_nodename.

When set the venus driver will create an OF entry in-memory consistent
with:

dec_nodename {
    compat = "video-decoder";
};

and/or

enc_nodename {
    compat = "video-encoder";
};

This will allow us to reuse the existing driver scheme of relying on compat
names maintaining compatibility with old dtb files.

dec_nodename can be "video-decoder" or "video0"
enc_nodename can be "video-encoder" or "video1"

This change relies on of_changeset() API as a result select OF_DYNAMIC will
be added to venus/Kconfig

Change-Id: Ia8c16125fffeebdf1720b2ad99c984506eb19eb4
Git-repo: https://gitlab.freedesktop.org/linux-media/media-committers.git
Git-commit: 687bfbba5a1cb15cee51519c210141ed1c4b0ec7

Tested-by: Renjiang Han <quic_renjiang@quicinc.com>
Change-Id: I8228605171a15d470024617931d1f2055a53c737
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Stanimir Varbanov <stanimir.k.varbanov@gmail.com>
Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>
Signed-off-by: Vikash Garodia <quic_vgarodia@quicinc.com>