msm: mhi_dev: Skip probe for non-PCIe boot configuration
With commit <735a0f425c98> ("msm: ep_pcie: Allow Clients to Query the
Link State during Non-PCIe Boot"), the EP PCIe driver sets the link
state to INVALID in the case of a non-PCIe boot. So, let's query the
link state, and if it is a non-PCIe boot (link state returned as
INVALID), then simply return failure without proceeding with any further
initialization.

Here, the MHI driver should depend on the EP PCIe driver (meaning the
MHI driver should probe only if the EP PCIe driver probes).
This change works only if this dependency chain is taken care of properly.

If Mhi driver proceeds with probe in non-PCIe boot, it's holding a
wakelock, which prevents system from suspend.

Change-Id: If4a737441f3a5b42256c9155140628496629eb49
Signed-off-by: Sumit Kumar <quic_sumk@quicinc.com>
Signed-off-by: Anvita T <quic_atadepal@quicinc.com>