PENDING: net: stmmac: enable SGMII loopback for DMA reset
enable SGMII Tx to Rx loopback bit as the RX clocks
from the AQR113C PHY are not available at the time
of the DMA reset. We can however extract the RX clock
from the internal SERDES block. Once the link is up,
we can revert to the previous state.

Change-Id: Ib3020c3a3f05f23b8ec47b396a3147273c380622
Signed-off-by: Sagar Cheluvegowda <quic_scheluve@quicinc.com>