FROMLIST: iommu/arm-smmu-qcom: NULL pointer check for driver data
The driver_data is NULL when qcom_adreno_smmu_init_context()
is called before the dev_set_drvdata() from the client driver
and is resulting in kernel crash.

So add a null pointer check to handle the scenario
where the client driver for the GPU SMMU device would
be setting the driver data after the smmu client device
probe is done and not necessarily before that. The function
qcom_adreno_smmu_init_context() assumes that the client
driver always set the driver data using dev_set_drvdata()
before the smmu client device probe, but this assumption
is not always true.

Change-Id: I78c2796cfd9678ce8b4f96a9a2692b8868eb62ea
Link: https://lore.kernel.org/linux-iommu/20230816225509.11070-1-quic_aprasann@quicinc.com
Signed-off-by: Aravind Vijayakumar <quic_aprasann@quicinc.com>
Signed-off-by: hanzh <quic_hanzh@quicinc.com>