phy: eusb: phy-qcom-snps-eusb2: Add support for SA8797P
This patch adds support for the SA8797P SoC to the phy-qcom-snps-eusb2
driver. The SA8797P platform allows multiple resources, such as clocks,
regulators, resets, TLMM, and interconnects, to be managed by firmware,
while the remaining responsibilities are handled by Linux.

The power domain, managed through the PM runtime framework, is utilized
for the activation and deactivation of these resources.

- Power domain "usb_core": Controls the GDSC, resets, and TLMM (GPIO).
- Power domain "usb_transfer": Controls clocks and interconnects.

In this variant, the two disjoint sets of resources are modeled against
the power domain suppliers, which can be turned on or off to achieve the
desired device states. Two modeled low power states, D1 and D3, are
supported.

transition: d0 <--> d3  :: both core and transfer suppliers
transition: d0 <--> d1  :: only transfer supplier

Change-Id: I258ed820d01cb605fb2f5958923f7eaf99fc0650
Signed-off-by: Hiren Gohel <quic_vhiren@quicinc.com>