msm: mhi_dev: Set mhi_dma_ready flag every time when DMA is ready
When the MHI DMA (IPA or ECPRI) is ready to service client requests, it
triggers the callback registered by the client driver (MHI in our case).
Within this callback, we set the mhi_dma_ready flag to indicate that the
MHI DMA is ready. However, this flag is only set if the MHI MMIO
initialization is not complete, which is causing an issue in the
following scenario:

	- Link is up, mhi_dma_provide_ops() is received and DMA callback
	  is registered, resulting in execution of MHI MMIO
	  initialization.
	- MHI MMIO initialization is complete.
	- If the callback is received from DMA here, mhi_dma_ready flag
	  is not set as MHI MMIO initialization is completed.

This results in skipping the mhi_dev_enable() call in subsequent MHI
reset handling, which results in device not polling for M0 and not
doing necessary initializations as part of mhi_dev_enable(). To resolve
this set mhi_dma_ready flag every time when DMA ready callback is
received.

Change-Id: I7ce286471270ac323a24b4297fc2372c6f2a79b5
Signed-off-by: Sai Chaitanya Kaveti <quic_skaveti@quicinc.com>