mmc: sdhci-msm: runtime_suspended flag needs to be reset
In sdhci_msm_runtime_resume function runtime_suspended flag
was not getting set to 0 when QoS is not defined.Due to this
IRQs were not getting processed for the host.

To fix this issue we are going to skip_qos block
instead of returning.

Change-Id: I114b31e9ea6495408e04992444788210ad54080d
Signed-off-by: Darshankumar Jagdishchandra Thakkar <quic_djagdish@quicinc.com>
Signed-off-by: Tarun Saxena <quic_tsaxena@quicinc.com>
(cherry picked from commit 5c67f03a576a0ce8c9d2939fc268271ff4b7ccd5)