usb: dwc3-msm: Set SMMU mode dynamically based on boot_config value
SMMU S1 translations needs to be disabled for USB in case of
M2 PCIE devices since IPA will be in bypass. Currently we use
overlay DT files for this purpose. But the new implementation
would identify the device type by reading the boot_config node,
and set iommu to bypass if it's M2 PCIE flavor.

Change-Id: I81cd7b6c65267fe304cac8022e80c3c8516735d3
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>