QCLINUX: arm64: defconfigs: Disable TBU support
Disable TBU support for gen4auto. This defconfig is only enabling
a context fault handler for adreno smmu. This context fault handler
is trying to get the physical address from the faulting iova from the
pgd pointed to by the pgtable ops. But GPU, with the per process
pagetable support, can change the pagetable ops depending upon the
GPU processes, so there could be edge cases where the faulting iova
will not be present in the pgd pointed to be the pgtable ops, if the
GPU have switched the pagetable ops and pagetable.

So disable this defconfig and use the default context fault handler
provided by the upstream kernel.

Change-Id: Ie669f79eabf982572ab6476305b846ec0d1485fe
Signed-off-by: Kaisheng Yang <quic_kaisheng@quicinc.com>