BACKPORT: usb: typec: Add support for Parade PS8830 Type-C Retimer
The Parade PS8830 is a USB4, DisplayPort and Thunderbolt 4 retimer,
controlled over I2C. It usually sits between a USB/DisplayPort PHY
and the Type-C connector, and provides orientation and altmode handling.

The boards that use this retimer are the ones featuring the Qualcomm
Snapdragon X Elite SoCs.

Add a driver with support for the following modes:
 - DisplayPort 4-lanes
 - DisplayPort 2-lanes + USB3
 - USB3

There is another variant of this retimer which is called PS8833. It seems
to be really similar to the PS8830, so future-proof this driver by
naming it ps883x.

Signed-off-by: Abel Vesa <abel.vesa@linaro.org>
Reviewed-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Link: https://lore.kernel.org/r/20250206-x1e80100-ps8830-v6-2-60b1e49cfa8d@linaro.org
Change-Id: Ied4674e7344be8504682bad6e9d6fbc29a1ba3f8
Git-commit: 257a087c8b5206e046048de6053fc8b3fa1af814
Git-repo: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
[rajashekar.kuruva@oss.qualcomm.com: merging with msm-specific changes]
Signed-off-by: Rajashekar Kuruva <rajashekar.kuruva@oss.qualcomm.com>