usb: dwc3: qcom-mp: Add support for shutdown for usb glue driver
For the multiVM targets, during guest VM reboot cycles,
as the shutdown path is not present, the clocks and GDSC are not voted
OFF. Hence the primary VM keeps the resources alive till next boot cycle.

So, during the warm boot, for the USB IP, the sequence of enablement of
resources becomes GDSC ON -> clcocks ON -> BCR assert -> BCR deassert.

This violates the recommended init sequence from the programming guide
which suggests the sequence should be GDSC ON -> BCR assert ->
BCR deassert -> clocks ON.

This voilation is causing improper init for USB. As an aftermath, the
xhci register space is ZEROed.

Disable the clocks and GDSC during shutdown path so that they will
be initing from scratch for all warm boot cycles.

Change-Id: Ic4a6c914a91df07ad79823fd0fe5be5ad05da698
Signed-off-by: Sriram Dash <quic_sriramd@quicinc.com>