soc: qcom: hab: Configure the DMA attribute of export memory to DMA_BIDIRECTIONAL
The current export memory dma map configuration is DMA_TO_DEVICE,
which means we expect the CPU to write and the device to read this memory.

However, we have a new application scenario where the device write
memory and the CPU reads memory. Therefore, we need to change the
export memory dma map configuration to DMA_BIDIRECTIONAL.

Change-Id: Ic48515cc84235a7fe2881118ed30c828fd394ad1
Signed-off-by: Zelong Ren <quic_zeloren@quicinc.com>