FROMLIST: arm64: dts: qcom: qcs8300-ride: Add firmware-name in BT node
The qcs8300-ride platform uses the QCA6698 Bluetooth chip. While the
QCA6698 shares the same IP core as the WCN6855, it has different RF
components and RAM sizes, requiring new firmware files. Use the
firmware-name property to specify the NVM and rampatch firmware to load.

Note: Only the firmware-name is being added, as the rest of the
changes have already been merged.

Link: https://lore.kernel.org/linux-arm-msm/20250211104421.1172892-2-quic_chejiang@quicinc.com/.

Change-Id: I05e14f614301827a7015b83209fa1f22d11b771c
Signed-off-by: Shuai Zhang <quic_shuaz@quicinc.com>