xfrm: Adjust IP4/IP6 output method for mixed tunnels
In some cases like tunnel mode with mixed inner/outer IP type
and packet offload, the chosen route IP type will be different
from the actual inner packet IP type. However, because of
the packet offload mode, the packet will be sent as is to
the network device driver.
To handle this case properly, we must adjust the IP output method,
rather than send with the outer header output method.

Change-Id: Id5ff1d939889744f434fc00ec7b287376f89d46b
Signed-off-by: Ilia Lin <quic_ilial@quicinc.com>