cfg80211: Don't reset ssid len and nlportid after a link removal
During a link removal from a MLD, stop ap is called
where nlportid is getting reset to 0. Which is wrong
as still the partner links and MLD will be up and running.
It causes the next consequent association to fail by rejecting M2
packet. This change fixes by resetting the nlportid only during
the last link removal/if all links are already
removed.

Same operation for ssid len as well.

Change-Id: Id250b96b716c9a5d0eae5a14c436293a6a921520
Signed-off-by: Kannan Saravanan <quic_kannsara@quicinc.com>