PENDING: usb: dwc3: qcom: Ensure VBUS_VALID is set after APPS Resume
During APPS Suspend, the clock controller unit disables all clocks
to the CPU and subsystems, causing the controller to enter the
Power-On Reset (POR) state. This results in the loss of
Qscratch programming.

Upon APPS Resume, the USB controller operating in device mode does
not have the VBUS_VALID signal set, which causes the PHYs to
transition into the suspend state. This behavior disrupts the
normal operation of the USB controller.

This commit addresses the issue by ensuring that the VBUS_VALID
signal is properly set for both High-Speed and SuperSpeed PHYs
after the system resumes from Deep Sleep. This is crucial
for maintaining the correct operational state of the USB
controller and preventing unintended suspensions of the PHYs.

Change-Id: Ic3fdfc366eb56104e79c7a727935dd1fe7cf8363
Signed-off-by: Sriram Dash <quic_sriramd@quicinc.com>