PENDING: usb: dwc3: qcom: Ensure PIPE_UTMI_CLK_SEL is proper after APPS Resume
During APPS Suspend, the clock controller unit disables all
clocks to the CPU and subsystems, causing the controller to
enter the Power-On Reset (POR) state. This results in the
loss of Qscratch programming.

Upon APPS Resume, the USB controller does not have the proper
PIPE_UTMI_CLK_SEL set. If only the High-Speed PHY is used for
a controller capable of SuperSpeed operations, the controller
will assume it should use the USB3 PHY and not the HS PHY as
configured before APPS Suspend. This behavior disrupts the
normal operation of the USB controller.

Address the issue by ensuring that the PIPE_UTMI_CLK_SEL is
properly set for the controller after the system resumes from
APPS Suspend. This is crucial for maintaining the correct
operational state of the USB controller and preventing
unintended suspensions of the PHYs.

Change-Id: I98e8d94652fd840fc8c1da35458636b5783eb2b6
Signed-off-by: Sriram Dash <quic_sriramd@quicinc.com>