clk: qcom: clk-alpha-pll: Add snapshot of slew PLL
Few alpha PLLs would require to be able to slew in the same VCO mode, which
means the PLL would be able to update the new frequency L value without
turning it off. But to be support this feature the PLL needs to calibrated
at the mid of the VCO range and then enabled.

Add support for dynamic update in which the frequency can be changed
without turning off the PLL. and also Update list_register operation
to read the pll offset defined in PLL register map to avoid mismatch
alpha PLL.

commit a7716b80271d ("clk: qcom: clk-alpha-pll: Add support to be
able to slew PLL").

Change-Id: I5bfa8d0751fb7798aa5250710391cb3383fee9df
Signed-off-by: Taniya Das <quic_tdas@quicinc.com>
Signed-off-by: Chintan Kothari <quic_ckothari@quicinc.com>
Signed-off-by: Asit Shah <quic_asitshah@quicinc.com>