BACKPORT: PCI/DOE: Rename Discovery Response Data Object Contents to type
PCIe r6.1 (which was published July 24) describes a "Vendor ID", a
"Data Object Type" and "Next Index" as the fields in the DOE
Discovery Response Data Object. The DOE driver currently uses
both the terms type and prot for the second element.

This patch renames all uses of the DOE Discovery Response Data Object
to use type as the second element of the object header, instead of
type/prot as it currently is.

Git-commit: f810d17762fba58b0b547c5fdd77a24e7aa1e939
Git-repo: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
Signed-off-by: Alistair Francis <alistair@alistair23.me>
Reviewed-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Link: https://lore.kernel.org/r/20250306075211.1855177-2-alistair@alistair23.me
Change-Id: Ia8691609bdcee050ab9ea9ea3d273efdabbe09e1
Signed-off-by: Nitesh Gupta <nitegupt@qti.qualcomm.com>