arm64: defconfig: sdxkova: Enable msm_performance
This commit enables the msm performance kernel module for sdxpinn
target. The perf service requires this module in order to vote for
the maximum and minimum CPU frequencies, among other things.
So by enabling this module, the perf service will be able to more
efficiently manage the CPU frequencies and improve system performance.

Change-Id: I0e8475d3380fc72e9cdb196765319257d62cefc6
Signed-off-by: Zhuo Fu <quic_zfu@quicinc.com>