tty: serial: qcom-geni-serial_msm: fix CTS control for UART in loopback mode
When running loopback with flow_control on and off,
uart_stop_tx is called due to a change in hw_stopped
from the serial core, resulting in no data transfer.
Consequently, the loopback remains in a wait state
for the transfer to complete.

To address this, set the TIOCM_CTS(sw-assisted CTS
flow control flag) control signal when the UART port
is in loopback mode. This ensures proper flow control.

Change-Id: I97f412ea6f9d10cca9863422b99d98f98cf8a953
Signed-off-by: Panicker Harish <quic_pharish@quicinc.com>
Signed-off-by: Maramaina Naresh <quic_mnaresh@quicinc.com>
Signed-off-by: Chiluka Harish <chiluka.harish@oss.qualcomm.com>