net: stmmac: Fix RX queue routing logic
In current Rx queue routing logic, we can route only
one type of packet to a queue even if we add multiple
routing entries in dtsi for that particular rx_queue
node. Fix this logic so that we can enable routing
of different type of packets to single queue based
on dtsi route entry.

Change-Id: Ia4a4653285a8c2fe5bd92bbb182f0c557e1c381e
Signed-off-by: Vatte Vishnuvardhan Reddy <quic_vishnuvr@quicinc.com>