PCI: Add change to support shutdown callback
This change adds a shutdown callback to PCIe ECAM driver to
ensure that it is properly shutdown in reboot/shutdown path.
If hardware is still accessing memory after SMMU is disabled
as part of SMMU shutdown callback in system reboot or shutdown
path, then it could lead to NoC or SMMU crashes.

Change-Id: Ib5b0d263464cee77f745257aac31f69949a65fd5
Signed-off-by: Mrinmay Sarkar <quic_msarkar@quicinc.com>