PENDING: arm64: dts: qcom: qcs9075-rb8: Enable displayport 0 and 1
The Qualcomm qcs9075 rb8 platform comes with 2 DisplayPort controllers
for each mdss. edp0 and edp1 correspond to the DP controllers of
mdss0, whereas edp2 and edp3 correspond to the DP controllers of
mdss1. This change enables only the DP controllers, DPTX0 and DPTX1
alongside their corresponding PHYs of mdss0, which have been validated.

Change-Id: Ic70c26221332f7a187f350f4c171e4d63020f640
Signed-off-by: Ritesh Kumar <quic_riteshk@quicinc.com>