soc: qcom: Add SCT table for neo le
Add SCT table for neo le target.

Here a slice id 69 has been resued for LLCC_SPAD which has been already been assigned to LLCC_EVA_3DR.
This is done keeping in note that there would not be LLCC_EVA_3DR for neo and SPAD would not be required on any other target.

Change-Id: I264ef585032b86c47b7bb896bb41fa67b60b9ee6
Signed-off-by: Amrit Anand <quic_amrianan@quicinc.com>