usb: phy: snps: Skip Phy powerdown for managed resources
For fw managed devices, if the genpd virtual devices
are put, then the control goes to firmware which
manages the resources. With no EUD SPOOF DISCONNECT,
the control is passed down to firmware, which is not
aware of the INIT or suspend states, or the role of
USB.

Hence, do not powerdown the PHY and let it be managed
via resources only. This way, we do not have to rely
on the role of DUT. and we can skip INIT for cable
disconnect and connect.

Change-Id: I8a43c62360c6e4f54341d3c8d4f411be28fa10da
Signed-off-by: Sriram Dash <quic_sriramd@quicinc.com>