usb: phy: snps: Optimize PHY Resource Handling with SCMI Channel Delays
During resource handling, when PHY initiates power on/off
instructions for genpd devices, the corresponding SCMI
channels and domains are enabled/disabled. The firmware
manages the initialization and deinitialization of the
associated SCMI channels/domains.

There is an approximate delay of 6 ms for each SCMI
transfer. To ensure proper resource enablement/disablement,
let the PHY device activities sleep for a while.

Change-Id: I1b1f4248fc8ed60617b556806ff703582401ca50
Signed-off-by: Sriram Dash <quic_sriramd@quicinc.com>