usb: dwc3-msm: Remove unnecessary dwc3 pointer init in dwc3_msm_resume
Remove the unnecessary platform_get_drvdata call and dwc pointer
initialization in the dwc3_msm_resume function.

The dwc pointer is not used in the function, making its initialization
redundant. Additionally, the variable u32 reg, which was previously
declared twice, is now declared only once at the beginning of the function.

This change simplifies the code structure and improves readability.

Change-Id: If29aa16a7393b4301dd55e5bd889ce7a4b58e69f
Signed-off-by: Rajashekar kuruva <quic_kuruva@quicinc.com>