usb: gadget: uvc: Initialize color matching descriptors for frame-based
The color_matching parameter has not been initialized for the frame-
based format, leading to a NULL pointer crash at __uvcg_fill_strm:
VCG_COLOR_MATCHING.
This patch adds the necessary initialization logic for color matching
in the uvcg_framebased_make function.

Additionally, it addresses a regression caused by the change
'https://lore.kernel.org/r/20230202114142.300858-7-
dan.scally@ideasonboard.com', which did not include bailout logic if a
 symlink is added for the frame-based format.

According to the change, "There's always a color matching descriptor
associated with the format, but without a symlink, it should only ever
be the default one. If it's not the default, there's already a symlink,
and we should bail out".

Change-Id: Ibb048c0d44dab78fdcdfc278fd2521fecb6378a4
Signed-off-by: Zihang Hu <quic_zihahu@quicinc.com>
Signed-off-by: Asit Shah <quic_asitshah@quicinc.com>