coresight-tmc-pcie: fix pcie mhi channel issue
This change enables PCIe mode support for the TMC driver and
introduces a sysfs node to allow selection of the pcie path.

Additionally, it addresses an issue where registering the mhi
state callback would fail if the mhi channel was already open.
The fix by checking the return value of the registration attempt,
if it returns -EEXIST, the driver will now manually invoke the
previously registered callback to ensure proper state handling.

Change-Id: I1f966667aaf4278f0b10faeda2b6ae29fa68235c
Signed-off-by: Yuanfang Zhang <quic_yuanfang@quicinc.com>