sched/walt: Remove rd->pd check after initialization
Currently in walt code we are waiting for perf domain creation and once
created we check again if perf domain are created properly or not, if
not we hit the BUG_ON condition. The check is redundant and this will
hit the BUG_ON condition even if rd has been rebuilt as old rd will be
freed. Remove this check to handle such scenarios.

Change-Id: I17e40a30607ed24fbd61981c87cda0d6ac7676c0
Signed-off-by: Ankit Sharma <anshar@qti.qualcomm.com>
(cherry picked from commit 3b9ba2729419db5ff5eae0ed375a0325cd8ae1a2)