FROMLIST: media: venus: pm_helpers: use opp-table for the frequency
The frequency value in the opp-table in the device tree and the freq_tbl
in the driver are the same.

Therefore, update pm_helpers.c to use the opp-table for frequency values
for the v4 core.
If getting data from the opp table fails, fall back to using the frequency
table.

Link: https://lore.kernel.org/linux-arm-msm/20250530-add-venus-for-qcs615-v8-1-c0092ac616d0@quicinc.com/
Reviewed-by: Vikash Garodia <quic_vgarodia@quicinc.com>
Change-Id: If6e18484595bc0957d4432ed282f42c86c42c3c5
Signed-off-by: Renjiang Han <quic_renjiang@quicinc.com>
Signed-off-by: haissun <haissun@qti.qualcomm.com>