serial: msm_geni_serial: Configurable vote for cpu, geni and ddr
The serial driver initializes resources with a fixed clock vote
for the CPU, Geni, and DDR. However, these values vary from target to
target.

To address this, read the clock vote values from the device tree
node and configure and initialize the resources accordingly. If
the resources are not explicitly mentioned in the Device tree node,
they will be initialized with default values.

Change-Id: Ic68a2b4ad02ff6de5f02db373103f3e4c4be5cc2
Signed-off-by: Prasanna S <quic_prass@quicinc.com>
Signed-off-by: Papanaboyina Bala Narayana <quic_papanabo@quicinc.com>