Revert "usb: dwc3-msm: Add GSI_EP_OP_PREPARE_TRBS under dwc spinlock"
This reverts commit bded8042198e5baaa2021cff0459e763060bd5ee.

Reason for revert: Don't assign memory using dma_alloc_coherent for
trb buffers via GFP_ATOMIC because GFP_ATOMIC limits the allocator
to memory that is immediately available and results in fail if
required more memory for buffer.
If memory is not fine than the logic used in this patch won't
work as patch used the spinlock which fails as memory allocation
is not ATOMIC and sleep.

Change-Id: Ic1cc3d0c92ba4ad1913a2cdb934e4680f598bbff
Signed-off-by: Manish Nagar <mnagar@qti.qualcomm.com>