drivers: iio: imu: Add wakelock support for IAM20680 sensor
Introduced a wakelock mechanism in the IAM20680 driver to ensure
reliable delivery of Wake-on-Motion (WoM) events to userspace.

When the sensor detects motion during system suspend, a wakelock is
acquired to prevent the system from suspending again immediately. This
ensures that the anymotion event is properly propagated to userspace,
giving it sufficient time to process the event and take appropriate
action. Without this wakelock, the system might resume briefly and
suspend again before userspace can respond, potentially missing the
event.

Change-Id: Ibbbd5c78191355361986ce786c1a61de0353d239
Signed-off-by: Puneet Yatnal <quic_puneety@quicinc.com>