drivers: input: imu: Add wakelock support for SMI230 sensor
1. Introduced a wakelock mechanism in the SMI230 driver to ensure
reliable delivery of Wake-on-Motion (WoM) events to userspace.

When the sensor detects motion during system suspend, a wakelock is
acquired to prevent the system from suspending again immediately. This
ensures that the anymotion event is properly propagated to userspace,
giving it sufficient time to process the event and take appropriate
action. Without this wakelock, the system might resume briefly and
suspend again before userspace can respond, potentially missing the
event.

2. Enabling SMI230 accel sensor during suspend to detect
anymotion feature.

Change-Id: I091b954a1ee24c07cce9bd17d81f67a35074cc51
Signed-off-by: Puneet Yatnal <quic_puneety@quicinc.com>