i2c: i2c-msm-geni: Configurable vote for cpu, geni and ddr
The I2C driver initializes resources with a fixed clock vote
for the CPU, Geni, and DDR. However, these values vary from target to
target.

To address this, read the clock vote values from the device tree
node and configure and initialize the resources accordingly. If
the resources are not explicitly mentioned in the device tree node,
they will be initialized with default values.

Change-Id: I313533d3ce1d4e96e0bc216d42bd34b00d323969
Signed-off-by: Papanaboyina Bala Narayana <quic_papanabo@quicinc.com>
Signed-off-by: Prasanna S <quic_prass@quicinc.com>