spi: spi-msm-geni: Configurable vote for cpu, geni and ddr
The SPI driver initializes resources with a fixed clock vote
for the CPU, Geni, and DDR. However, these values vary from target to
target.

To address this, read the clock vote values from the device tree file and
configure and initialize the resources accordingly. If the resources are
not explicitly mentioned in the device tree, they will be initialized with
default values.

Change-Id: I418f65200d66536fff53a7170e3f942d7d430fd7
Signed-off-by: Mehul Raninga <quic_mraninga@quicinc.com>
Signed-off-by: Anil Veshala Veshala <quic_aveshala@quicinc.com>
Signed-off-by: Prasanna S <quic_prass@quicinc.com>