scsi: ufs: qcom: Deregister panic notifier cleanly
As a part of ufs_qcom_init(), we conditionally register to panic notifier.
This registration is cleared as a part of ufs_qcom_remove(), but not in
ufs_qcom_exit(). This is causing the panic notifier list entries to get
corrupted when ufshcd_init() fails and control reaches out_disable.

Therefore, deregister panic notifier in exit path as well to avoid
notifier list corruption.

Change-Id: Ia1a8b42aa031f5f92cab9a7cb42f3422a819b42a
Signed-off-by: Anjana Hari <quic_ahari@quicinc.com>