usb: dwc3-msm: Use level low trigger while enabling DP/DM irqs
Commit d0b87b4d1f55 ("usb: dwc3-msm: Add additional configuration
while enabling DP/DM irqs") configures DP and DM interrupts as
'EDGE_FALLING | LEVEL_LOW', but PDC only accepts either level
or edge configuration, but not combination of both. Because of
this, the enablement of wakeup intterrupts fail. Hence use only
LEVEL_LOW as resume triggers for DP (in HS) and DM (in LS/FS).

Change-Id: I25494b79e6b7c2f73f63ded431771c0aee26e133
Signed-off-by: Prashanth K <prashanth.k@oss.qualcomm.com>