scsi: ufs: qcom: Set spm level to 5 during restore
In cases where UFS is operating on Guest VM, spm level will
be intially at default level 3 and the hibernation image has
this spm level value. But when platform suspend is invoked for
suspend to disk usecase, the whole system will go to power down
state and a complete reset and restore of UFS subsystem is needed.
During this step, in QNX based targets, UFS clock rates are not set
to the required value, but are reset to 19.2MHz from QNX host driver
which will lead in UFS operational failures. To avoid this scenario,
we scale down clocks just before entering resume path, conditionally
when spm level is 5.

However, the snapshot image is having spm level 3 which will stop
the control from reaching the conditional clock handling.

Set spm level to 5 during restore state to ensure
successful reset and restore operation during hibernation.

Change-Id: I0f4de1cfb939a111794a22efd683628106d1a7ea
Signed-off-by: Anjana Hari <quic_ahari@quicinc.com>