BACKPORT: UPSTREAM: arm_scmi: move boiler plate code into the get info functions
This code to check whether the selector is valid and if the item has
already been recorded in the array can be moved to the
scmi_pinctrl_get_function_info() type functions.  That way it's in
one place instead of duplicated in each of the callers.

I removed the check for if "pi->nr_groups == 0" because if that were the
case then "selector >= pi->nr_groups" would already be true.  It already
was not checked for pins so this makes things a bit more uniform.

I also removed the check for if (!pin) since pin is an offset into the
middle of an array and can't be NULL.

Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>
Reviewed-by: Cristian Marussi <cristian.marussi@arm.com>
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>
Link: https://lore.kernel.org/all/287b5302f583e3535d50617ec3b0856e38253171.1761576798.git.dan.carpenter@linaro.org/
Git-commit: a3c46c82d8a2075dfe2a2af1431486c8b128e93c
Git-repo: https://git.kernel.org/pub/scm/linux/kernel/git/sudeep.holla/linux.git/
Change-Id: I36e23d57cc64d614abcaee2ba3d893c62f33c9ba
Signed-off-by: Zong Jiang <quic_zongjian@quicinc.com>
(cherry picked from commit cac609d77236ae7a47bbabba19deff9af92c1c98)