usb: dwc3: dwc3-msm: Allow runtime suspend only after pm_resume
Currently when the device is pm_suspended due to bus suspend,
if we get a host initiated resume (in form of dp/dm/ss irq),
we set resume_pending flag and call resume_work(). We try to
do runtime_resume() from resume_work(), which will fail since
the device is pm_suspended, but then we set resume_pending to
false incorrectly and wait for pm_resume(). When pm_resume()
kicks in, we fail to call dwc3_msm_resume() as resume_pending
flag is cleared. This causes bus resume to fail.

Fix this by checking resume_pending flag only after pm_resume()
is completed.

Change-Id: I46107f880a2911194f0b008ad16f2d602f2db558
Signed-off-by: Prashanth K <prashanth.k@oss.qualcomm.com>