FROMLIST: clk: qcom: gcc: Update the SDCC clock to use shared_floor_ops
gcc_sdcc2_apps_clk_src: rcg didn't update its configuration" during
boot. This happens due to the floor_ops tries to update the rcg
configuration even if the clock is not enabled.
The shared_floor_ops ensures that the new parent configuration is
cached in the parked_cfg in the case where the clock is off.

Ensure to use the ops for the other SDCC clock instances as well.

Change-Id: Ic680e1d0488a30892276a16f9fe85c1ad6067a56
Link: https://lore.kernel.org/all/20250804-sdcc_rcg2_shared_ops-v1-1-41f989e8cbb1@oss.qualcomm.com/
Fixes: 39d6dcf67fe9 ("clk: qcom: gcc: Add support for QCS615 GCC clocks")
Signed-off-by: Taniya Das <taniya.das@oss.qualcomm.com>
Signed-off-by: Dongfang Zhao <dongfang.zhao@oss.qualcomm.com>