PENDING: usb: misc: eud: Specify initial role during probe
In cases where the role switch provider is missing, the initial
role would be marked NONE and never DEVICE. When eud is enabled on
such platforms, the enable_eud call would bail out since role is not
device. Fix the initial role to device for the following reasons:

1. Primary controller is the only one using eud. If there is a role
switch provider and DUT is booting up in host mode, device would
get a role switch.

2. If controllers dr mode is host, then EUD is not for use on that
controller.

3. If dr_mode is peripheral, it is safe to mark the role as DEVICE.

Change-Id: I8d53c9c51a4c392ec2944c7d0253874b7beb61dc
Signed-off-by: Krishna Kurapati <krishna.kurapati@oss.qualcomm.com>