QCLINUX: pci: quirks: Disable L1.1 and L1.2 for USB hub attach
When ASPM L1.1/L1.2 is enabled on the upstream link of QPS615
switch,and Renesas uPD720201 USB 3.0 Controller is connected
to one of the switch downstream ports, then uPD dies down and
stops responding to host commands.

Disabling l1 sub-states on the upstream link in uPD attach
case helps mitigate this issue.

Change-Id: I427fc5faa16168097dcb1225b7f44efde8b6115b
Signed-off-by: Sushrut Shree Trivedi <quic_sushruts@quicinc.com>