generic_vm_defconfig: Disable some config options
In order to save some memory on the VM side, sort and disable
some of the options that do not seem to be needeed - including
firmware loader, ipv6 support, various networking options and
hardware drivers, filesystems, android kabi struct paddings etc.

With this change, the free memory on the VM increases with ~864k
on the debug build and with ~852k on perf build. Below is the
change on the vmlinux:

   text    data     bss     dec     hex filename
8165568 2845552  337592 11348712         ad2ae8 ./org.vmlinux
   text    data     bss     dec     hex filename
7457883 2702224  326256 10486363         a0025b ./new.vmlinux

Change-Id: I3b37577a3aa74a7b219fee61a0036cdd2043da67
Signed-off-by: Georgi Djakov <quic_c_gdjako@quicinc.com>