gh_virtio_pci: Disable the suspend/resume routines for virtual PCI devices
The resume routines of virtual PCI devices re-enumerates all of the devices
on the bus. During re-enumeration, the VM exits to CrosVM multiple times
increasing latency of resume. There can be delays in scheduling CrosVM
threads that are emulating the PCI bus. This adds to the delay on resuming
the VM from suspend. During this time, VM could be in the noirq context
and the time to services IRQs might significantly rise. In a virtual
environment as suspend/resume of pci doesn't add much value, disable them
to reduce the IRQ latency.

Change-Id: I75a01f4663a4ce54dda31492b9d601f883378da7
Signed-off-by: Prakruthi Deepak Heragu <prakruthi.heragu@oss.qualcomm.com>