soc: qcom: gic_intr_routing: Handle defective cpu case
There are certain scenarios where logical CPU is not equal to the
physical cpu in HLOS. In this case, correct mpidr of the cpu that is
referred by logical cpu number is to be used.

Handle this case by reading cpuid_mpidr for the cpus that are available.
Map this value with the logical CPU number and use it directly.

Change-Id: I58b8639ff913eb794d5b334592283bab3c261c33
Signed-off-by: Lakshmi Sowjanya D <laksd@qti.qualcomm.com>
Signed-off-by: Rohit Jadhav <rbjadhav@qti.qualcomm.com>