sched/walt: add checks for proxy execution
Currently during walt load balancing we are not skipping to move those
tasks which are waiting on a mutex or the tasks which are currently
chosen by the sched to run but it is in a waiting state and instead the
"owner" task possibly is running on the cpu. Add these checks to
essentially not move those during load balancing.

Change-Id: I9baa1a79095ed107e20b4bc8bc635a6868ef4a9f
Signed-off-by: Ankit Sharma <ankit.sharma@oss.qualcomm.com>
Signed-off-by: Ashay Jaiswal <ashay.jaiswal@oss.qualcomm.com>