serial: msm_geni_serial: Use runtime PM usage count during shutdown
Updated the msm_geni_serial_shutdown() logic to use the device's
runtime PM usage count instead of the internal ioctl_count to
determine whether to power on the device.

This change ensures better alignment with the kernel's power
management framework and avoids the scenario to depend on the
ioctl count as it is leading to the incorrect device usage count.

Change-Id: I8ea28fb793640f76ffc6ce702794dcce91fa08ab
Signed-off-by: Krishna Chaithanya Reddy G <quic_kgangapu@quicinc.com>
Signed-off-by: Viken Dadhaniya <viken.dadhaniya@oss.qualcomm.com>