soc: qcom: Add kmsg dump for QTVM on watchdog bark
Add kernel message dumping to shared memory when watchdog bark occurs.

Add a call to kmsg_dump() with KMSG_DUMP_OOPS in the watchdog bark
handler. This triggers the gh_panic handler on the QTVM side, which dumps
the kernel log buffer to shared memory before the watchdog bite occurs.

Change-Id: I245f37061452cf8c34f4b7601298ffded11ccde2
Signed-off-by: Mukesh Pilaniya <mukesh.pilaniya@oss.qualcomm.com>