sched/walt: add framework for inst overhead study
Introduce framework to track number of instructions consumed by individual
tracehooks. The framework handles preemption and re-entry from
tracehooks using stack based tracking. The stack is per cpu instead of
per thread to keep the overheads to a minimum.

Not having thread based stack, could lead to incorrect behaviour.
For example, a thread running on cpuX calls tracehook A and B. The stack
for cpuX would have A and B in that order expecting an exit from B
followed by A. However if that task moves to a different cpu, cpuX
continues to maintain A and B in it stack. Any new event on cpuX will
appear as if it is interrupting tracehook B and will account the
instructions until its entry to B. This leads to unusully high inst count
on B. This can also happen when a cpuX switches tasks.

Also, a migrated task to a new cpu, could try to pop from the stack
and would land with no entries. This situation is handled "gracefully"
where such instructions are accounted towards the INVALID bucket.

We could disable preemption at the tracehook entry to avoid the
migration aspect but that introduces overhead and changes system behaviour.

Instead the onus of detecting such a situation is put on user.
One can read the debug/walt/trace_hook_inst_cnt back to back and
if entries in INVALID bucket are seen along with percpu stack in
staying same state - it could be a migration situation.
The user is advised to issue a reset - by writing anything to
debug/walt/trace_hook_inst_cnt - to reset the stack amongst other counters
and try again.

Such a stuck stack happens more often with WAKEUP_NEW, SCHED_FORK_INIT and
CPU_CGROUP_ATTACH as these are amongst the tracehooks called in a
switchable context.

In the future we could look in to tracking stack at the task level,
thereby restoring  the stack  back when the task switches in on a cpu,
or simply not tracking the tracehooks that are called in sleepable
context.

Thankfully most of the tracehooks and functions (walt_irq_work, utra) are
called under rq locks - they won't switch tasks. Governor is usually called
with rq lock held or in an irq context so it too won't switch.
The issue is only present with tracehooks that
can be called in switchable context.

Change-Id: If4805ef37fb302ded476c58539d36f06b85e2b5d
Signed-off-by: Ashay Jaiswal <quic_ashayj@quicinc.com>
Signed-off-by: Abhijeet Dharmapurikar <quic_adharmap@quicinc.com>
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>